added CLOCK_SPEED parameter to NES object - python version 0.3.4

+++ include/universal/cpu.h
@@ -188,7 +188,8 @@ class CPU {
         uint16_t Y;
         uint16_t A;
         uint16_t S;
-        uint16_t PC;
+        uint32_t PC;
+        
         int clock_speeds[3] = {1790000,2680000,3580000};
 
 };
